[RAG] Refactor imports and enhance document processing
- Simplified import statements in `rag.py` and `rag.py` for better organization and readability.
- Updated the `process_document` function in `rag.py` to use an asynchronous client for Supabase, improving performance.
- Enhanced metadata handling in the document processing pipeline to ensure proper updates even when metadata is initially None.
- Cleaned up the import statements in `pgmq_init.py` and `rollback_pgmq.py` for consistency.

Questions restantes : Are there further optimizations needed for the document processing workflow?
