+++ backend/scripts/pgmq_init.py
@@ -9,7 +9,7 @@ import logging
 
 from sqlalchemy import text
 
-from app.database.session import engine
+from app.core.db import engine
 
 logger = logging.getLogger(__name__)
