+++ backend/scripts/rollback_pgmq.py
@@ -9,7 +9,7 @@ import sys
 
 from sqlalchemy import text
 
-from app.database.session import engine
+from app.core.db import engine
 
 logger = logging.getLogger(__name__)
